Change the extensions spec in the cabal file.
The instances in the library are not really undecidable, but the current ghc
version requires the -fallow-undecidable-instances flags (older versions did
not require this flag).  Unfortunatelly, hackage does not allow uploads
of packages with this flag---instead it insists that the option has to be
specified with the UndecidableInstances language extension.  This is wrong,
but until hackage & GHC are fixed we change the cabal specification.